+++ src/main/kotlin/no/nav/familie/ef/iverksett/brev/frittstående/DistribuerKarakterutskriftBrevTask.kt
@@ -97,7 +97,7 @@ class DistribuerKarakterutskriftBrevTask(
     }
 
     override fun onCompletion(task: Task) {
-        taskService.save(Task(OppdaterOppgaveTask.TYPE, task.payload, task.metadata))
+        taskService.save(Task(OppdaterKarakterinnhentingOppgaveTask.TYPE, task.payload, task.metadata))
     }
 
     companion object {
