+++ src/main/kotlin/no/nav/familie/ef/iverksett/brev/frittstående/JournalførKarakterutskriftBrevTask.kt
@@ -62,6 +62,6 @@ class JournalførKarakterutskriftBrevTask(
 
     companion object {
 
-        const val TYPE = "journalførKarakterutskriftBrev"
+        const val TYPE = "JournalførKarakterutskriftBrevTask"
     }
 }
